+++ web-frontend/modules/database/plugin.js
@@ -33,6 +33,7 @@ import {
   CSVImporterType,
   PasteImporterType,
   XMLImporterType,
+  JSONImporterType,
 } from '@baserow/modules/database/importerTypes'
 import { APITokenSettingsType } from '@baserow/modules/database/settingsTypes'
 
@@ -80,6 +81,7 @@ export default ({ store, app }) => {
   app.$registry.register('importer', new CSVImporterType())
   app.$registry.register('importer', new PasteImporterType())
   app.$registry.register('importer', new XMLImporterType())
+  app.$registry.register('importer', new JSONImporterType())
   app.$registry.register('settings', new APITokenSettingsType())
 
   registerRealtimeEvents(app.$realtime)